# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: products
    description: Product management operations
  - name: movements
    description: Stock movement operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters and pagination
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: lowStock
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, -name, quantity, -quantity, price, -price, createdAt, -createdAt]
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProductById
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: productId
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string, enum: [IN, OUT] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementListResponse"
    post:
      operationId: createMovement
      tags: [movements]
      summary: Record stock movement (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovementInput"
      responses:
        "201":
          description: Movement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movement"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          description: Optional structured error details
      required: [message]
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, viewer] }
      required: [id, name, email, role]
    RegisterInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        role: { type: string, enum: [admin, viewer] }
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    AuthSuccess:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        quantity: { type: integer }
        lowStockThreshold: { type: integer }
        isLowStock: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, sku, category, price, quantity, lowStockThreshold, isLowStock]
    ProductInput:
      type: object
      properties:
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number, minimum: 0 }
        quantity: { type: integer, minimum: 0 }
        lowStockThreshold: { type: integer, minimum: 0 }
      required: [name, sku, category, price]
    ProductUpdateInput:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [items, page, limit, total, totalPages]
    ProductRef:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
      required: [id, name, sku]
    MovementInput:
      type: object
      properties:
        productId: { type: string }
        type: { type: string, enum: [IN, OUT] }
        quantity: { type: integer, minimum: 1 }
        note: { type: string }
      required: [productId, type, quantity]
    Movement:
      type: object
      properties:
        id: { type: string }
        product:
          $ref: "#/components/schemas/ProductRef"
        type: { type: string, enum: [IN, OUT] }
        quantity: { type: integer }
        note: { type: string }
        userId: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, product, type, quantity, userId]
    MovementListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [items, page, limit, total, totalPages]
    DashboardSummary:
      type: object
      properties:
        totalProducts: { type: integer }
        lowStockCount: { type: integer }
        lowStockItems:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        totalStockValue: { type: number }
        totalUnits: { type: integer }
        categories:
          type: array
          items:
            type: object
            properties:
              category: { type: string }
              count: { type: integer }
        recentMovements:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
      required: [totalProducts, lowStockCount, lowStockItems, totalStockValue, totalUnits, categories, recentMovements]